Resolved incompatible Type of event ID, TODO, change all of them top long, also in db

+++ backend/src/main/java/com/joinmatch/backend/dto/EventResponseDto.java
@@ -6,7 +6,7 @@ import java.time.LocalDate;
 import java.time.LocalDateTime;
 
 public record EventResponseDto(
-        Long eventId,
+        Integer eventId,
         String eventName,
         Integer numberOfParticipants,
         BigDecimal cost,
